Major refactoring (#59)
* Started refactoring into multiple files. ST seems to be working again.

Separated the components into a directory and different files.

* Move png functions

* Create tab folder and create main tab file

* Rename component files

* Refactor B0shim tab

* Use __ST_DIR__ and __CURR_DIR__

* Refactor B1 tab

* Refactor dicom_to_nifti tab

* Refactor fieldmap tab

* Refactor mask tab

* Remove unused variables

* Add option for auto masking

* Rename dynamic tab

* Fix newlines at the end of files
